requirements.txt
google-genai==1.71.0
docling==2.85.0
langchain-text-splitters==1.1.1